added docstrings to /Home/ components
