+++ frontend/src/components/Home/Antivirals.js
@@ -1,10 +1,11 @@
-// FIXME: antivirals do not appear in the left-hand summary pane (<DisplayedParameters />) until the component rerenders!
-// there's a workaround implemented for NPIs by changing a state variable in the parent component (<HomeView />) when an intervention is added
-// changing state re-renders <HomeView /> and <DisplayedParameters /> updated dynamically
+// FIXME: antivirals do not appear in the left-hand summary pane (<DisplayedParameters />) until the component
+// rerenders! there's a workaround implemented for NPIs by changing a state variable in the parent component
+// (<Home />) when an intervention is added.
+// changing state re-renders <Home /> and <DisplayedParameters /> updates dynamically
 
 // a form for gathering user specification for antiviral interventions, which are saved to state and localStorage
 // clicking the "Antivirals" button in the <Interventions /> dropdown will render this component
-// NOTE: specifying antiviral interventions does not currently affect disease spread!
+// NOTE: specifying antiviral interventions does not currently affect the simulation!
 
 import React, { useState, useEffect } from 'react';
 import toggletip from  '../images/toggletip.svg';
