+++ frontend/src/components/Home/NPI.js
@@ -1,3 +1,5 @@
+// a form for collecting user specifications for non-pharmaceutical interventions (or NPIs)
+// clicking on the Non-Pharmaceutical button in the <Interventions /> dropdown will render this component
 import React, { useState, useEffect } from 'react';
 import Select from 'react-select';
 import toggletip from  '../images/toggletip.svg';
@@ -5,7 +7,7 @@ import './InitialCases.css'; // Import the CSS file for styling
 
 const texasMapping = require('../../data/texasMapping.json'); // Import the Texas mapping JSON
 
-const NonPharmaceutical = ({ counties, onSubmit }) => {
+const NPIForm = ({ counties, onSubmit }) => {
   const [nonpharmaName, setNonpharmaName] = useState(localStorage.getItem('nonpharma_name') || 'School Closures');
   const [nonpharmaDay, setNonpharmaDay] = useState(40);
   const [nonpharmaDuration, setNonpharmaDuration] = useState(10);
@@ -299,4 +301,4 @@ const NonPharmaceutical = ({ counties, onSubmit }) => {
   );
 };
 
-export default NonPharmaceutical;
+export default NPIForm;
