+++ frontend/src/components/Home/PlayPauseButton.js
@@ -1,3 +1,6 @@
+// a button responsible for starting and stopping the simulation
+// PlayPauseButton only changes the STATE of whether the simulation is running, the logic for
+// controlling the simulation occurs in the parent component, <Home />
 import React from 'react';
 import './PlayPauseButton.css';
 import play_button from '../images/play_button.png';
