+++ frontend/src/components/Home/SetScenario.js
@@ -1,5 +1,6 @@
+// renders the Set Scenario dropdown menu on the left panel of Home view
 import React, { useState } from 'react';
-import AddInitialCases from './InitialCases';
+import InitialCases from './InitialCases';
 import SetManually from './DiseaseParameters';
 import './InitialCases.css'; // Import the CSS file for styling
 import { createPortal } from 'react-dom';
@@ -22,7 +23,7 @@ const SetScenario = ({ counties, onSave, scenarioChange, casesChange }) => {
   const closeSetManually = () => setIsSetManuallyOpen(false);
 
   const handleSave = () => {
-    scenarioChange();    // trigger HomeView.js to rerender with updated scenario info
+    scenarioChange();    // trigger Home.js to rerender with updated scenario info
     onSave(); // Trigger the save action in the parent component
     setShowDropdown(false); // Close the dropdown after saving
   };
@@ -67,7 +68,7 @@ const SetScenario = ({ counties, onSave, scenarioChange, casesChange }) => {
               <div className="modal-contents" style={{ backgroundColor: 'white' }} onClick={(e) => e.stopPropagation()}>
                 <span className="modal-close" onClick={closeInitialCases}>×</span>
                 <h2>Add Initial Cases</h2>
-                <AddInitialCases 
+                <InitialCases 
                   onClose={closeInitialCases} 
                   counties={counties} 
                   onSubmit={(data) => { console.log('Initial Cases:', data); handleSave(); }}
