+++ frontend/src/components/Home/SpreadMapPercent.js
@@ -1,3 +1,5 @@
+// one of two main map components that tracks the infections per county as a percentage of that county's population
+// this component is invoked in <Home /> and renders in the top-middle of Home view
 import React, { useState, useEffect, useRef } from 'react';
 import { MapContainer, TileLayer, GeoJSON, useMap } from 'react-leaflet';
 import L, { transformation } from 'leaflet';
@@ -94,7 +96,7 @@ const Legend = () => {
 };
 
 
-const InfectedMapPercent = ({ eventData, currentIndex }) => {
+const SpreadMapPercent = ({ eventData, currentIndex }) => {
   const [countyData, setCountyData] = useState([]);
   const mapRef = useRef();
 
@@ -179,4 +181,4 @@ const onEachCounty = (feature, layer) => {
   );
 };
 
-export default InfectedMapPercent;
+export default SpreadMapPercent;
